hid buttons on job form based on logged in user.

+++ app/views/temp_jobs/show.blade.php
@@ -12,14 +12,16 @@
     <h5>Date Needed: {{{ $job->required_date }}} </h5>
     <h5>Created By: {{ $job->creator->first_name }} {{ $job->creator->last_name }}</h5>
 
-    <!-- For giver to delete job -->
-    {{ Form::open(array('action' => array('JobsController@destroy', $job->id), 'method' => 'DELETE')) }}
-        <button type="submit" class="btn btn-sm btn-danger">Delete</button>
-    {{ Form::close() }}
-
-    <!-- For helper to apply for job -->
-    {{ Form::open(array('action' => array('JobsController@apply', $job->id), 'method' => 'POST')) }}
-        <button type="submit" class="btn btn-sm btn-succces">Apply</button>
-    {{ Form::close() }}
+    @if(Auth::user()->is_helper == FALSE)
+        <!-- For giver to delete job -->
+        {{ Form::open(array('action' => array('JobsController@destroy', $job->id), 'method' => 'DELETE')) }}
+            <button type="submit" class="btn btn-sm btn-danger">Delete</button>
+        {{ Form::close() }}
+    @elseif(Auth::user()->is_helper == TRUE)
+        <!-- For helper to apply for job -->
+        {{ Form::open(array('action' => array('JobsController@apply', $job->id), 'method' => 'POST')) }}
+            <button type="submit" class="btn btn-sm btn-succces">Apply</button>
+        {{ Form::close() }}
+    @endif
 </body>
 </html>
